requirements.txt
PyOpenGL==3.1.6
PyOpenGL_accelerate==3.1.6
glfw==2.5.5
Pillow==9.3.0
numpy==1.23.4